{config,snap}state: add new refresh.metered=force option and flip default
This PR adds a new option to refresh.metered: "force". This option
will force automatic refreshes even on metered connections. The
default (when unset) is now to not do automatic refreshes.

However if the maximum time that refreshes can be defered
(60 days) is over the refresh is forced even if the network is
metered (this has always been the case).