ifacestate: fix hang when retrying content providers
The old code was a bit too naive when checking for content-providers
in auto-connect. It was assuming that when a content provider is
installed in the same change the task that needs the content provider
waiting for that is ok. However this only works if there are no
wait relations between the tasks.

This PR adds a check that ensures that doAutoConnect() does not
hang forever if it finds a content provider install in its
wait chain. In this case the install will just continued and
the content provider auto-connect will DTRT and connect the
interfaces.

This fixes the annoying hang on cosmic that is described in
LP: #1776622.