cmd/snap-update-ns: enforce trespassing checks
This patch enables and enforces the trespassing checks. The rules as as
follows. Writes to /tmp (which is managed by snap-confine)
/var/lib/snap, and /snap/$SNAP_NAME are unrestricted. That is, they may
potentially write to the host filesystem. All other writes are
restricted and are internally replaced with a writable mimic, as if we
had encountered a read-only directory where we wish to create a new file
or directory for content bind mount.

The main function constructs the Assumptions object an pass it down all
the way to the function that performs mount changes. There is quite a
lot of argument passing to eventually pass the restrictions object to
filesystem operations such as MkDirAll, MkFileAll or MkSymlinkAll.

Performed mount changes are tracked so that we can keep track of the
exact set of tmpfs filesystems that were mounted by snapd. This requires
no new outside state so we are compatible with existing mount namespaces
and are safe to upgrade to.

There are ample new unit tests and, most importantly, the layout
integration test no longer needs to correct for the bug!

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>