tests/snapctl-disable-services-install-hook: don't redundantly do check
This is all unnecessary, because we really only care about the exit code and the script is already being run under "-e" so it will exit if snapctl fails.
Additionally, if it does exit with non-zero exit code, snap install will already output stuff from stdout/stderror so we will be able to see it failed.