many: allow using statx by default
The statx(2) system call was introduced with the Linux kernel 4.11 and
has slowly gained adoption in userspace since then. It offers multiple
improvements over the older stat interface.

The system call is now used in the wild by Qt but since the default
template doesn't allow it applications get permission denied errors
instead. This is undesired as statx is equally safe and just as stat
itself, should be allowed by default. The manual page states:

   To access a file's status, no permissions are required on the file
   itself, but in the case of statx() with a pathname, execute (search)
   permission is required on all of the directories in pathname that
   lead to the file.

Unfortunately due to a bug in libseccomp-golang we cannot easily know
that a given system call name (such as "statx") is unknown to seccomp.
As a result we cannot detect when snapd was built with older seccomp and
golang-seccomp and thus cannot support statx correctly even if the
kernel does.

This patch makes the system call available on Ubuntu Cosmic and other
distributions where seccomp 2.3.3, or newer, is available.

Fixes: https://github.com/snapcore/core18/issues/4
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>