client, cmd/snap: on !linux, exit when the client tries to Do something
Before this change snap client on non-linux would exit with a warning
when trying to instantiate a client.Client object inside a command. As
we need to move towards instantiating client before running the
commands parser and passing that client in to the command, this
approach proves problematic. This change makes it so that only when
trying to talk to the network does the non-linux client bail out.