many: return real snap name in API response
This patch partially reverts the work that was done for mapping "core"
or "snapd" snap names to the "system" nickname. The nickname is still
accepted on input, the "core" word is still used in the on-disk state
file but the real snap name (either "core" or "snapd") is now being
returned from API responses, unchanged.

This reverts the API break when a "system" snap was returned in API
responses but that snap was not installed in the system (in fact, system
cannot be installed).

The client side is tweaked to understand how to abbreviate the "core",
"snapd" or "system" snap names.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>