overlord/state: return latest LastAdded time in WarningsSummary
Before this change, state's WarningsSummary returned the current
timestamp together with the warnings count. This was fine for then
calling Okay, but useless as a way to filter out the WARNING
footer. This change makes it so that the timestamp returned is the
time the latest showable warning was added, instead.