Moving core-snap-refresh-on-core test from main to nested suite
This is done because it is set as manual in the main suite and it is
failing when the core is refreshed. By moving the test to the nested
suite, it is easier to run it and control the reboots which are
currently failing.

Currently it is failing because after the revert (from edge to stable),
the installed core is the stable one and the tracking is edge, which is
wrong and makes the test fail.

See the error log:
https://paste.ubuntu.com/p/vw7DsjxZhk/