overlord/configstate: export experimental flags to C
Experimental flags are available in snapd but, before this patch, not in
the C world of snap-confine and state-agnostic world of snap-update-ns.

This patch fixes this by exporting core.experimental.* flags to
/var/lib/snapd/facts/experimental. Facts are pieces of truth as
determined by snapd, available to C and Go where reading the state or
re-implementing a piece of Go functionality in C is undesirable.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>