vendor, cmd/snap: refactor to accommodate the new less buggy go-flags
go-flags on master has some fixes to manpage generation that are
interesting for us. In particular, the paragraphs following a command
with only hidden options were not dedented properly, and included a
spurious "Usage:" line, making it very messy and hard to find the
following command:

```
   help
       Show help about a command

       The help command displays information about snap commands.

       Usage: snap [OPTIONS] help [help-OPTIONS]

          info
              Show detailed information about snaps
```

would now render properly as

```
   help
       Show help about a command

       The help command displays information about snap commands.

   info
       Show detailed information about snaps
```

as well as some the alignment of options in help output no longer
leaving space for hidden options.

However, another bug fixed by this revision is that the
`PassAfterNonOption` flag now performs as advertised, meaning that
`snap install foo --classic` and `snap install --classic foo` are
different, if that flag is used. This is bad for us because we want
the fexibility it used to provide, but we also need this behaviour for
`snap run`.

That is, everywhere except in `snap run` we want options that come
after arguments to be treated as options, but in `snap run`,
anything that comes after the snap name argument we want to pass
through.

The `PassDoubleDash` flag would let us do that, by simply including a
`"--"` before the snap name. Unfortunately if we switch to requiring
this, it will break people who might be running `snap run` by hand and
dependend on the old behaviour.

This commit drops our use of `PassAfterNonOption`, and tweaks
`os.Args` to add a `"--"` before the snapname in `snap run` calls, if
the user hasn't added it themself.

HTH, HAND.