cmd/snap: for "snap run", add PassAfterNonOption
After sleeping on it, I realised this approach was better: instead of
figuring out whether a `"--"` is needed and adding it (an approach
that is error-prone, and in fact the previous implementation would
fail if one tried e.g. `snap run --hook configure snapapp` instead of
`snap run --hook=configure`), we check if we're trying to do `snap
run` and just add the needed flags option back.