cmd/snap-update-ns: better detection of snapd-made tmpfs
This patch extends the conservative logic for determining if a directory
can be safely written to because as a snapd-made tmpfs. The initial
logic was only accepting the mount point that corresponds to an entry
in the current mount profile, that is, direct proof correlation between
a tmpfs in the system and a tmpfs mount directive as given by snapd.

This works fine for cases where we mount tmpfs over a directory, e.g.
/bin and subsequently create files or directories as entries directly
underneath. It also works in cases where we create a brand new
directory, for example /bin/foo/ and then populate it with files or more
directories. This second case is not directly allowed by the trespassing
tmpfs detector. It only works because the detector is switched off as
soon as, while creating an object such as /bin/foo/bar we noticed that
we created foo as a new directory. This proves that foo was not an
existing bind mount and that it was created on a tmpfs substrate and
thus inhabits the same tmpfs.

This logic was too restrictive though, when /bin/foo/ already existed
and we attempted to create /bin/foo/froz. In such case the old logic
would allow writes to /bin (because it matches a tmpfs mount point) but
would not allow the write in /bin/foo to be attempted because it was not
inspecting /bin/foo nor trying to correlate that to /bin itself in any
way.

With the new change as we discover that /bin is a tmpfs that snapd
instructed to mount (perhaps sometime in the past when old copy of snapd
ran) we collect the major:minor and store it for the duration of the
snap-update-ns process.

When we subsequently determine trespassing safety of /bin/foo we keep
allowing writes for as long as the major:minor numbers can be traced to
a tmpfs that we visited and verified in the past. This removes the
conservative restriction while keeping the tmpfs verification strict. It
is strict in the sense that unknown tmpfs that is mounted by another
party (e.g. part of the classic distribution partitioning scheme) is not
considered safe to write. This way we still guarantee that writes do not
leak to the initial mount namespace.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>