overlord/snapshotstate: store the SnapID in snapshot, block restore if changed
This change makes it so that the snap ID is stored in the snapshot,
and if on restore that ID is non-empty, and there is a snap with the
snapshot's snap's name installed that has a non-empty snap ID, and
these two IDs are different, the restore is fails.

This feature was requested at the Brussels sprint, as a minimal
prerequisite to renames (and merges). This then needs to grow to
handle the rename, but at least for now it will do no damage.