cmd/snap-update-ns: remove empty placeholders used for mounting
This patch fixes an edge case where a mount namespace cannot be updated
because of leftover empty placeholder files or directories that
interfere with the subsequent operation, e.g. construction of a symlink
or directory (while a placeholder file exists).

When a mount change is undone, empty files and empty directories that
inhabit file systems other than squashfs are removed. This is imperfect
because we don't know if we had actually created a file but since we
measure it as empty the compromise is acceptable.

The spread test that was introduced prior to the fix is updated to show
the correct operation of revert now.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>