wrappers/services.go: handle disabled sockets, timers during start svc task
Currently, `systemctl is-enabled SERVICE.sock` for a service with sockets always returns "static", and so the previous logic would ignore whether a socket/service had been disabled or not. Instead, we should check if a service has sockets, and if it does then don't use is-enabled on the service itself, and instead check the sockets.
Also note that currently one can't ever disable a service with sockets/timers using `snap stop --disable` or `snapctl stop --disable`, changes are needed for that.