snap, client, daemon, store: use and expose "media" more (#5906)
* snap, client, daemon, store: use and expose "media" more

Upcoming GUI work needs access to the "media" property from the
store. I'd hoped to be able to delay this until we'd dropped v1 (and
thus screenshots), but they need it RSN, and dropping v1 is looking
like a six-month thing. This exposes "media" straight to the client,
and populates screenshots and icon url from it instead of requesting
them individually.

* daemon, snap, store: Drop IconURL and Screenshots from snap.Info

Instead of `info.IconURL`, you now do `info.Media.IconURL()`; instead
of `info.Screenshots` you now do `info.Media.Screenshots()`. Thanks
@pedronis for pointing this possibility out; much nicer.

Also I droped the Screenshot() method on the individual MediaInfo as
it was weird.

* daemon: keep vet happy