cmd: rename ns_group to mount_ns
The concept of namespace groups was designed in the hope that there
would be multiple types of namespaces used together for the purpose of a
given snap. Since only mount namespaces are used and we need to handle
per-snap mount namespaces along with the new per-snap and per-user mount
namespaces, being clear in terminology is important.

As such rename "ns_group" with "mount_ns".

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>