systemd: extend Status() to work for socket and timer units
In wake of extending reporting of service statuses in `snap services` output we
will need to ask for the status of timer and socket units. Those do not have the
Type property. Extend systemd.Status() to work for timers and sockets. While at
it, tweak the field names of returned struct to closer reflect that those are
units (and not necessarily services).

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>