cmd: separate helper process startup code
The helper process is being decoupled from the rest of the mount
namespace initialization process to allow it to exist even if the mount
namespace of the snap is correct and up to date. This is needed to allow
it to still live and capture the per-user version of the mount
namespace.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>