cmd: separate namespace inspection from helper life-cycle
This patch allows the mount namespace capture helper process to have
separate life-cycle from the call to the function
sc_create_or_join_mount_ns. The said function is reduced to
sc_join_preserved_ns that can return ESRCH if preserved namespace did
not exist. This allows snap-confine to start the helper, unshare,
prepare the mount namespace, capture it and then ask the helper to exit.

In subsequent patches this ability will be exploited to keep the helper
alive long enough to also capture the per-user mount namespace.

With this reshuffling snap-confine knows exactly when the namespace must
be initialized so the need for the sc_should_populate_mount_ns is gone,
as is the function and the associated state.

Some of the comments that were rendered stale in earlier patches are
corrected.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>