interfaces/system-key: add parser mtime and only discover features on write
Because snap run will generate a system-key to compare with what is on
the system, the previous commit to add parser features support caused
apparmor_parser --preprocess to be called on every snap run invocation,
which is not allowed by the policy, not performant and more importantly,
not intended or needed.

The patch adds apparmor-parser-mtime to the system-key, adds it to
generateKey(), moves discovering the parser features to when the
system-key file is written to disk and omits verifying that the parser
features are the same in SystemKeyMismatch. In this manner, only a
stat() call to obtain the mtime is needed by snap run (fast and no
policy changes), snapd will regenerate the system-key when it discovers
the mtime changes and the parser features are always on disk for profile
generation.