daemon: support `snap logs snapd` to get snapd logs
This PR adds support for `snap logs snapd` to show the snapd.service
log. With this PR the experience with an installed snapd snap becomes
more consistent. The other benefit is that this way users of the
snapd API (like a management snap) can access the snapd logs for
debugging and error detection.