tests: create dummy per-user mount namespace using stdlib, not libmount
This patch fixes an issue that only manifests itself on Ubuntu 14.04
where /etc/mtab is a real file and not a symbolic link to /proc/mtab.
On such system mount(1) which uses libmount internally, keeps track
of all the operations by editing the mtab file. Sadly, our hackery inside
snap-confine and snap-discard-ns is totally not editing mtab. This causes
confusion when umount(1) is asked to detach a mount namespace file.

The problem is that at that time the file is already unmounted, thanks
to updated snap-discard-ns. Because the test had atificially created the
.mnt file with touch and mount(1) there is now a stale mount entry in mtab.
In the cleanup section of each test (actually the prepare section but
conceptually it is the cleanup after previous test) we, via snapd.postrm
or snap-mgmt.sh unmount and remove all the preserved mount namespaces.
When umount(1) is given a file that is not a mount point it goes to
search mtab, treating it as a mount _source_ (not target) and proceeds
to umount the target of whatever is found.

To avoid this issue, since we are not using libmount in snap-discard-ns
or snap-confine, simply avoid using libmount driven mount(1) in the test
as well.

This patch implements a thin python wrapper around mount and uses it
for the construction of the fake per-user mount namespace.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>