tests/main/snap-service-after-before-install: verify after/before in snap install
Add spread tests for verification of after/before handling during `snap
install`. To be able to observe the whole process in detail, the after/before
snap had to be modified. Switching to `daemon: notify` allows to have a
controlled time gap between the process leaving inactive and entering active
states.

Further, even before the fix was implemented, the processes could have been
started in the right order. This was internally caused how the application map
of the snap.Info was iterated over in a semi-random fashion. For this reason,
the actual test case is repeated 10 times. With just 3 service applications, it
should be possible to rule out the test passing by accident.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>