testutil, cmd/snap: introduce and use testutil.EqualsWrapped and fly
This introduces a family of checkers to `testutil`, which are useful
for when you want to check for a string in output that can be wrapped
to fit the terminal, or aligned as part of a table. E.g. if you want
to check for the third line in

```
Name   Version                      Rev   Tracking  Publisher   Notes
bare   1.0                          5     edge      canonical✓  base
black  18.6b4                       1     edge      sergiusens  devmode
core   16-2.36~pre2+git967.54466bd  5780  edge      canonical✓  core
```

then you can use

```go
Check(theLine, EqualsPadded, "black 18.6b4 1 edge sergiusens devmode")
```

Similarily, to check that the output of installing a `devmode` snap
without `--devmode`,

```
error: The publisher of snap "eggg" has indicated that they do not consider this
       revision to be of production quality and that it is only meant for
       development or testing at this point. As a consequence this snap will not
       refresh automatically and may perform arbitrary system changes outside of
       the security sandbox snaps are generally confined to, which may put your
       system at risk.

       If you understand and want to proceed repeat the command including
       --devmode; if instead you want to install the snap forcing it into strict
       confinement repeat the command including --jailmode.
```

if you wanted to check that it says "only meant for development",
which might be on the same line or not depending on terminal width,
you'd simply say

```go
Check(output, ContainsWrapped, "only meant for development")
```

HTH, HAND!