overlord/snapstate, snap, wrappers: start services in the right order during install
* snap: refactor after/before validation to separate helper for sorting services

Refactor the code for validation fo after/before application cycles into a
separate helper function that returns a sorted list.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>

* wrappers: start the services one by one, preserving caller provided order

Starting the services in a single `systemctl start a b c` call does not actually
create a single transaction, but instead produces a number of parallel jobs.
Ensure that services are started in the order provided by the caller by callling
`systemctl start` separately for each service.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>

* overlord/snapstate: ensure the services are properly ordered when starting

Ensure that services are started in the right order during snap installation.

Fixes: https://bugs.launchpad.net/snapd/+bug/1796125

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>

* snap: reference the original paper about topological sort

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>

* wrapper: properly pass a fixed list of services in unit tests

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>