overlord/standby: refactor a chunk more
This should do the trick. But I might've gone a little overboard.

Without this change, StandbyOpinions had a bunch of attributes that
were changed for synchronization with no locking. I think that's
problematic, so with this change it's a lot more like a tomb.

I also made it so Stop would wait for the goroutine to exit, as
@bboozzoo suggested. Tomblike, also, that.

I *also* also reuse the timer instead of creating a new one every
iteration. As this loop runs forever, this might make the garbage
collector less busy.

I refrained from changing the protocol so CanStandby returns
yes/no/never, with yes/no being the current True/False, and never
stopping the whole thing. Maybe next time.