cmd/snap-update-ns: parse the -u <uid> command line option
This patch extends the C command line parser to support the -u <uid>
option. The option implies we are being called by snapd to update the
mount namespace of a given user. It implies we should jump to the
appropriate namespace ourselves, use per-user mount profile and perform
our task.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>