strutil: let MatchCounter work with a nil regexp
Some concerns were raised about usig MatchCounter with a regexp to
match whole lines. This addresses that.

From the benchmarks:

    BenchmarkNil         3000   480553 ns/op    64354 B/op        802 allocs/op
    BenchmarkNilAll      2000   619675 ns/op   260776 B/op       2980 allocs/op
    BenchmarkNilEquiv     500  3673992 ns/op   953872 B/op       6544 allocs/op
    BenchmarkNilEquivAll  500  3837003 ns/op  1150292 B/op       8722 allocs/op

the two first use the new code, the first one saving just 3 lines, the
second one any number of them.

The second two use the old code (with a regexp of `(?m).+` which is
equivalent to the `nil` case). The new code is both faster and more
efficient, as expected.