daemon: spool sideloaded snap into blob dir
Before this change a sideloaded snap would get written out to /tmp
before being copied into /var/lib/snapd/snaps. As these are often
separate devices this means sideloading a snap would be unnecessarily
slow, and use up a potentially large amount of /tmp (which might be
small and in-memory, on core).

This change makes it so that the snap is instead written to a
temporary (and hidden) file in /var/lib/snapd/snaps directly.