packaging/opensuse: stop using golang-packaging
This patch switches the openSUSE packaging to stop using go-packaging.
The state of golang in openSUSE is not perfect. The existing helpers are
very cumbersome, verbose and don't provide useful features. In cases
like passing special linker flags the helpers require us to work around
them.

I've been in contact with Go maintainers in openSUSE and they recommend
to just use go directly. The only real value of the packaging helpers as
they exist today is to pass "-buildmode pie". In this patch I did just
that, switching to pure go for building.

The patch refers to "indigo" that I'm also working on, as a better and
less weird packaging helper for distributions. I used some the same
directory structure during the build process that Indigo uses directly
and I plan to switch to Indigo 0.3 or 0.4, as time permits. This will
simply replace "GOPATH=... go build -buildmode pie" with "indigo build"
and remove some of the setup boilerplate.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>