cmd/snap-confine: always put the snap process under a device cgroup
When no devices are tagged for the given snap, snap-confine will not put the
current process under a device cgroup. This may lead to the snap having access
to more devices than allowed by its connected interfaces.

Example case when this happens is input devices, where AppArmor is not
expressive enough and selective access is implemented using a device cgroup.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>