packaging/opensuse: add missing build-dep on systemd
This took some time to debug.

In absence of systemd.pc we guess the value of systemd system generator
directory (this is hard-coded in cmd/configure.ac). If the build
environment doesn't implicitly have systemd installed then this no
longer matches the RPM systemd macros that name said directories.

This would subtly break builds on openSUSE Leap 42.3 and Leap 15.0.

The patch includes extra build dependencies, additional sanity check to
ensure that RPM macros and pkg-config are in sync, default definition of
the macros (for Leap 42.3) and lastly the use of said macros instead of
hardcoded paths.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>