packaging/fedora: support 'rpmbuild --with apparmor'
This intentionally does not change the defaults for fedora/amazon/etc
but allows building with 'rpmbuild --with apparmor -ba ...'. For now it
requires libapparmor-devel, libapparmor1 and apparmor-parser rpms are
installed (which do not currently exist on these distros).

For full mediation, dbus 1.10 should be available and compiled with
'--with-apparmor' along with an apparmor-enabled kernel.