snap-confine: write 'apparmor' to display_lsm when available
LSM stacking allows systems with multiple configured LSMs. On these
systems, all LSMs are consulted for access with one set as the 'primary
LSM' that applications normally see (configurable via the kernel command
line). Eg, a Fedora system with stacking patches in the kernel and
AppArmor userspace available might have both SELinux and AppArmor
enabled, but SELinux would be the primary LSM, whereas on Ubuntu with
SELinux enabled, AppArmor would be the primary LSM. So that existing
userspace applications do not have to change, all applications by
default see the primary LSM, but applications can adjust the display LSM
so that it and all child processes see an alternate LSM.

Adjust snap-confine to unconditionally set the display LSM to 'apparmor'
when AppArmor is in use so that snaps don't have to worry about when
AppArmor isn't the primary LSM.

Note, the LSM stacking patches are not yet upstream but are available in
Ubuntu kernels as a technology preview since Ubuntu 17.10. As such, the
behavior does not change non-Ubuntu systems or default Ubuntu systems
(since the display LSM is already "apparmor"). This patch is meant to
allow experimenting with the feature.