interfaces/opengl: add additional accesses for cuda (#5696)
* interfaces/builtin/opengl.go: add additional cuda bits to opengl

* interfaces/builtin/opengl.go: remove duplicated nvidia /dev perms

* interfaces/builtin/opengl: drop SECCOMP rules enabling mknod

The SECCOMP rule was only added to allow CUDA libraries to mknod
/dev/nvidia-uvm-tools. However, there is some concerns that we may be allowing
too much, as the simple rules we use cannot verify if only allowed path is used.

Loading the driver and creating /dev/nvidia-uvm-tools manully before running
sample programs using CUDA seems to work. As a result, the patch drops the
SECCOMP rules and we'll instead direct people to the right workaround.