snap: make Epoch's MarshalJSON not simplify
Before this change, snap.Epoch's marshaller tried to use the
simplified notation when possible, i.e. it'd write `"1*"` instead of
`{"read":[0,1],"write":[1]}`.

While we still want to do this when printing it to the user, the store
has dropped support for the more compact representation (if it ever
had it), which means we'd need a different marshaller there. Rather
than that, this just drops the smarts from the marshaller. String()
still calls simplify() so that path is unchanged (and we explicitly
don't serialise to YAML).