interfaces: export gpio pin in AppArmorConnectedPlug
By exporting the gpio pin in the apparmor connected plug code we
avoid the implicit dependency between the apparmor and the systemd
security backends.

The layering is not ideal but by doing it this way we avoid
complicating the generic interfaces code that so far has no
concept of dependencies between security backends. If we ever
add such a concept we can tweak this the gpioInterface by modeling
that for a gpio connection we always run the "slot systemd backend
first".

This should also allow us to get rid of the order when adding
the backends and we could potentially run all security setup in
parallel (this PR is not doing any of this though :)