cmd/snap-confine: capture initialized per-user mount ns
When the experimental feature with persistent per-user mount namespaces
is enabled then a new mount namespace is created in /run/snapd/ns,
alongside $SNAP_NAME.mnt we now get $SNAP_NAME.$UID.mnt.

Note that per-user mount namespaces are not supported for the root user.
As a design decision root user is synonymous with the per-snap mount
namespace.

Preserved per-user mount namespaces are not inspected to detect
staleness. Instead the per-snap mount namespace, if stale, is discarded
alongside the mount namespaces of all the users of that snap on the
system, by calling snap-discard-ns.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>