cmd/snap-confine: handle mounted shared /run/snapd/ns
The code that handles special sharing mode of /run/snapd/ns did cope
with /run/snapd/ns being a private mount point and not being a mount
point at all. When it was a mount point but one with shared mount event
propagation all kinds of hell would break loose. This patch fixes this.

The existing unit tests were nixed and replaced with a much more
readable and comprehensive spread test.

Fixes: https://bugs.launchpad.net/snapd/+bug/1803542
Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>