many: add snap debug refresh-catalogs
This command can be used to force the refresh of the "catalogs", namely
some of the store side meta-data collection that is cached locally for
performance.

The command is used to improve one of the existing tests that fails on
timeout if the store responds with a 403, like it does now.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>