tests: make tinyproxy support systemd notify
The tinyproxy has a bit of a startup delay (python init and all that).
This means that sometimes in tests the proxy is not fully ready when
the test tries to talk to it. This PR makes the proxy a "notify"
service that only notifies after the proxy is ready.