cmd/snap: make 'snap warnings' output yamlish
The first time we tried to use `snap warnings` we found we wanted to
be far more verbose than the current output sensibly allowed. We need
to be verbose because we don't want to say what went wrong, we want to
say how to fix it -- which needs more space.

So this PR changes `snap warnings` output to be yaml-ish.

It means `snap warnings` grows a `--unicode` option (for en-dash), but
as there's no need for color this pr also splits colorMixin.

In a followup I can drop Unicode from colorMixin and make them really
separate (it's probably a little confusing right now). This can be
done as a refactor free of user-visible side effects.