overlord/snapstate, etc: rename "extra info" to "store info"
@zyga rightly pointed out that the "extra"' in "extra info" didn't
convey any meaning (thanks!).

I've renamed things, I think in a way that makes things clearer:

The cache directory is now `/var/cache/snapd/store`. It's called
`dirs.SnapStoreInfoCacheDir`.

The data structure that is dumped into a per-snap-instance file in
that directory is called `StoreInfo`.

The action of adding a snap's store info to the cache is called
`CacheStoreInfo`.

The action of taking the info from the cache and adding it to a
snap.Info is called `AttachStoreInfo`.

The action of removing the info of a snap from the cache is called
`DeleteStoreInfoCache`.

Additionally, on creating the cache dir a README is written to it
explaining what it's about.

Lastly, a snap not having a cached info is no longer considered
newsworthy; no error is returned, nothing is logged.