overlord/snapstate: on refresh, check new rev can read current
This adds checks to install and update paths so that you can't
accidentally (or on purpose) install a snap that can't read the
previous snap's data, as indicated by their epochs.

The case for refreshing from the store is there just to catch
mistakes, as the store should never hand out a revision with an epoch
that can't read the current one.

The case for InstallPath (used by both `snap try` and `snap install
--dangerous ./some-file.snap`) is more "real". There are spread tests
for install / try combos.

This adds both an early check (deciding with the metadata already
known before the changes are scheduled), and a late check (deciding
with the metadata obtained from the snap itself, during the mount
task).