tests: remove test-snapd-with-configure on restore
The tests/main/ubuntu-core-gadget-config-defaults test, when executed,
leaves the mount namespace of the test-snapd-with-configure snap. For
some tests that is not a problem so the mistake was left unnoticed for a
long time. For a new test I was writing it does matter though hence the
chase for leftovers in the test suite.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>