tests: restore in restore, not prepare
The test suite is not doing any restore logic in per-test-case restore
section. Instead that is done in the front of the per-test-case prepare
section. The reason for that is unclear but it makes no apparent sense
on a quick inspection.

Putting actual restore in the restore section will open the path to
adding safeguards that system state before prepare is measurably the
same as after restore, that is, that test doesn't affect the system.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>