run-checks: discard test good/bad banner
The endmsg function prints a banner that shows if tests passed or
failed. There are two problems with the banner:

It rarely formats correctly, making error logs show wrapped mess.

It is printed at the end of the execution of several separate tests
so it is possible to end up with a log that looks roughly like this:

- good stuff
- some error stuff
- lots of good stuff (pages and pages)
- +1 message

So we are getting:
 - big banner saying it's good
 - error code saying it is bad

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>