cmd/snap-confine: refactor calling snapd tools into helper module
The module allows opening tool by name, opening the two tools we
are known to open (snap-{update,discard}-ns) and calling a function
by open descriptor, with apparmor confinement if desired.

There's a small skew that tool.c includes snap-confine/apparmor-support.h
but that's a low price to pay for this cleanup.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>