overlord/snapstate: use file timestamp to initialize timer
Without this change, the catalog refresh wass done every time snapd
started. Usually not too bad, but in tests or dev it could get
onerous.

This change makes it so that the time of the last catalog refresh is
initialized from the timestamp of the sections file, which is a
sensible lower bound for the last time the catalog refresh was done
(unless the system clock is broken -- against which we protect by
ignoring times that are in the future).